Redesign of calculation code
Instead of doing calculations in controller app now uses a factory and
an object to do all the calculations. The factory even attaches the
calculations to sessionStorage upon factory return.
Calling this factory calculate function and pass it the variable it
needs to do calculation in the controller function for calculation page